States moved to own class
